refactor(rubygems): Merge `InternalRubyGemsDatasource` into main class
